Add end-to-end integration test
This commit sets up an integration test that verifies that the app makes the correct calls to AWS.

This is a bit of a unorthodox setup.
There's only one test and it starts the apps and watches the calls to AWS services.
This test uses AWS SDK mock to verify that the app make the expected invocations.

There's a conditional added to the app to stop it after running for a second in the test environment.

NOTE that this code is not intended to make it to production.
This simply provides a testing harness that can be used to verify that things work while we modularize the implementation and make the move from ECS to cloud.gov task-runners.
This should be phased out by more robust unit and integration tests before being deployed.